Fix refresh token endpoint

+++ backend/src/main/java/com/Adapter/icare/Configurations/CustomUserDetails.java
@@ -41,7 +41,7 @@ public class CustomUserDetails implements UserDetails {
         return user.getPassword();
     }
 
-    public Integer getId() { return user.getId(); }
+    public Long getId() { return user.getId(); }
 
     @Override
     public String getUsername() {
